[Patient Intake Sample] moves addLanguage function to intake-utils
